+++ programs/albus-swap/src/utils.rs
@@ -29,8 +29,59 @@
 use crate::errors::SwapError;
 use crate::state::TokenSwap;
 use anchor_lang::prelude::*;
+use solana_program::entrypoint::ProgramResult;
+use solana_program::program::invoke;
 use solana_program::program_memory::sol_memcmp;
 use solana_program::pubkey::PUBKEY_BYTES;
+use solana_program::system_instruction;
+
+/// Resize an account using realloc, lifted from Solana Cookbook.
+#[inline(always)]
+pub fn realloc_account<'a>(
+    target_account: &AccountInfo<'a>,
+    funding_account: &AccountInfo<'a>,
+    system_program: &AccountInfo<'a>,
+    new_size: usize,
+    refund: bool,
+) -> ProgramResult {
+    let rent = Rent::get()?;
+    let old_minimum_balance = rent.minimum_balance(target_account.data_len());
+    let new_minimum_balance = rent.minimum_balance(new_size);
+    let lamports_diff = new_minimum_balance.abs_diff(old_minimum_balance);
+
+    if new_minimum_balance > old_minimum_balance {
+        invoke(
+            &system_instruction::transfer(funding_account.key, target_account.key, lamports_diff),
+            &[
+                funding_account.clone(),
+                target_account.clone(),
+                system_program.clone(),
+            ],
+        )?;
+    } else if refund {
+        transfer_lamports_from_pdas(target_account, funding_account, lamports_diff)?;
+    }
+
+    target_account.realloc(new_size, false)
+}
+
+pub fn transfer_lamports_from_pdas<'a>(
+    from: &AccountInfo<'a>,
+    to: &AccountInfo<'a>,
+    lamports: u64,
+) -> ProgramResult {
+    **from.lamports.borrow_mut() = from
+        .lamports()
+        .checked_sub(lamports)
+        .ok_or(ProgramError::ArithmeticOverflow)?;
+
+    **to.lamports.borrow_mut() = to
+        .lamports()
+        .checked_add(lamports)
+        .ok_or(ProgramError::ArithmeticOverflow)?;
+
+    Ok(())
+}
 
 pub fn to_u128(val: u64) -> std::result::Result<u128, SwapError> {
     val.try_into().map_err(|_| SwapError::ConversionFailure)