+++ packages/albus-sdk/CHANGELOG.md
@@ -1,8 +1,47 @@
 # Changelog
 
+## [0.3.0-next.4](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.3.0-next.3...@albus-finance/sdk-v0.3.0-next.4) (2024-01-09)
+
+
+### 🌟 Features
+
+* **sdk:** support create and prove request with `txBuilder` ([3f1051b](https://github.com/mfactory-lab/albus/commit/3f1051bbe7b8dbc5244c73a0018ff1bdc5677b80))
+
+## [0.3.0-next.3](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.3.0-next.2...@albus-finance/sdk-v0.3.0-next.3) (2024-01-05)
+
+
+### 🐞 Bug Fixes
+
+* **sdk:** add `mint` for credential listener callback ([852911e](https://github.com/mfactory-lab/albus/commit/852911ed86975283852f0867ffd2f0a6c5d08d1c))
+
+## [0.3.0-next.2](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.3.0-next.1...@albus-finance/sdk-v0.3.0-next.2) (2024-01-05)
+
+
+### 🐞 Bug Fixes
+
+* **sdk:** fixed credential `addListener` ([e3d0651](https://github.com/mfactory-lab/albus/commit/e3d065165eab8fad5b9c2f53b97c19c64cfaa27e))
+
+## [0.3.0-next.1](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.2.5-next.6...@albus-finance/sdk-v0.3.0-next.1) (2024-01-05)
+
+
+### 🌟 Features
+
+* **sdk:** add credential listener ([4ee83f1](https://github.com/mfactory-lab/albus/commit/4ee83f1865fbe91582ff86b73dc1d3ab8021f247))
+
+
+### 🐞 Bug Fixes
+
+* **sdk:** add `deleteByAddr` for `policyManager` ([309e0d2](https://github.com/mfactory-lab/albus/commit/309e0d206a9a9139edd077cb234175127456cde2))
+* **sdk:** version ([3d92bde](https://github.com/mfactory-lab/albus/commit/3d92bdea1e4ba7d4c819ca8b266309764f032862))
+
+
+### ♻️ Refactors
+
+* **sdk:** code cleanup ([9fb02df](https://github.com/mfactory-lab/albus/commit/9fb02df138c11f29d9f5115f0fe8951494187430))
+
 ## [0.2.6](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.2.5-next.6...@albus-finance/sdk-v0.2.6) (2023-12-19)
 
-## [v0.2.4](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.2.5-next.6...@albus-finance/sdk-v0.2.6) (2023-12-04)
+## [v0.2.4](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.2.5-next.6...@albus-finance/sdk-v0.3.0-next.1) (2023-12-04)
 
 ## [v0.2.2](https://github.com/mfactory-lab/albus/compare/@albus-finance/sdk-v0.2.2-next.21...@albus-finance/sdk-v0.2.4) (2023-11-17)